requirements.txt
Django==2.2.24
djangorestframework==3.11.2
requests==2.26.0
